Refactor response handling.. and more
The parsing of the diagnostic command responses was moved to the command handler (in configureMQTT)

The main loop (with the timerInterval) rewritten to only send MQTT commands to request diagnostics

Changed the log level of mqtt config message to debug (they tend to fill logs with repetitive data)

If a 'diagnostics' command is sent with no options, default to requesting all supported diagnostic options for the vehicle

If the command handler doesn't understand the response data (all except location and diagnostics), publish a state topic containing the JSON returned by the server.  For example, unlockDoor will publish to /homeassistant/VIN/sensor/unlockdoor/state and JSON message that indicates a success/failure status. Be aware that this is the servers response to the command, and not the vehicles response.