Better HA auto-discovery
Most of the commands now publish an MQTT button discovery topic (allowing them to be
used without scripting.)

Location now publishes a proper MQTT device_tracker for discovery that works without
any oddness.

When a command succeeds (other than location/diagnostics) or fails (all), it will
publish an MQTT Event (but MQTT events are broken in HA) giving lots of information
about the success or failure.  While the events are broken, it can still be used
as a state trigger.  (Trigger on the event's 'Timestamp' attribute changing. A simple
action for this is to send a notification of '{{ trigger.to_state.attributes.friendlyMessage }}'

Probably some more things changed I forgot about.  I'm not done yet.